indicators/requirements.txt
numpy>=1.26
psycopg2-binary>=2.9
fastapi>=0.104
uvicorn>=0.24
